SeekCursor refactoring ()
This PR contains refactoring of a piece in SeekCursor that hopefully makes it more readable and easier to understand (at least for me :) ). This part is focused on initial tree traversal.

Intermediate commits are left to follow process of refactoring.
Main points:

traverseDownToCorrectLevel refactored to separate traversal down the tree and initialization of scan
avoid usage of pointerId for two different purposes
some common method extractions and redundant code removals
fix SeekCursorTestBase so it can run concurrently

+++ community/index/src/main/java/org/neo4j/index/internal/gbptree/GenerationSafePointer.java
@@ -47,8 +47,8 @@ import org.neo4j.io.pagecache.PageCursor;
  * </ol>
  */
 class GenerationSafePointer {
-    private static final int EMPTY_POINTER = 0;
-    static final int EMPTY_GENERATION = 0;
+    private static final long EMPTY_POINTER = 0L;
+    static final long EMPTY_GENERATION = 0L;
 
     static final long MIN_GENERATION = 1L;
     static final long FIRST_STABLE_GENERATION = MIN_GENERATION;
